requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
python-magic==0.4.27

# Development (optional)
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1